add Window, set staticruntime "Off" in premake
code-wise just added the windowing functionality for mswindows, but most
of the work was in wrangling the linker and header files.

Had to move the prebuilt header include in Core.h to after the #define
directives, was getting a sort of circular header include situation
where the prebuilt header included the Log.h file and Log.H needed
Core.h to have hit those #defines before including it ... Dunno if
there's a better way to fix that issue. (maybe have the prebuilt header
include core before including other Sauropod libraries would do it?)

had an issue where GLFW & SauropodEngine projects in visual studios were
setting the runtime library setting to different values (/MDd and /MTd I
think) that caused a linker error which I resolved by setting them both
to `staticruntime "Off"` in premake, dunno if that's the right setting
for them to have, but they both have to have the same setting or it
won't link properly (see
https://stackoverflow.com/questions/3007312/resolving-lnk4098-defaultlib-msvcrt-conflicts-with ).